requirements.txt
python-telegram-bot==21.5
apscheduler==3.10.4
groq==0.11.0
httpx==0.27.2